ElasticSearchProject/requirements.txt
asgiref==3.7.2
certifi==2024.2.2
Django==5.0.2
django-elasticsearch-dsl==8.0
django-elasticsearch-dsl-drf==0.22.5
django-nine==0.2.7
djangorestframework==3.14.0
elastic-transport==8.12.0
elasticsearch==8.12.1
elasticsearch-dsl==8.12.0
Faker==23.3.0
packaging==23.2
python-dateutil==2.9.0
pytz==2024.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.10.0
urllib3==2.2.1
